refactor: move utils to feature
