+++ frontend/src/app/admin/layout.tsx
@@ -1,6 +1,6 @@
 import "./globals.css";
-import Sidebar from "@/components/sidebar/Sidebar";
-import Header from "@/components/header/Header";
+import Sidebar from "@/components/layout/Sidebar";
+import Header from "@/components/layout/Header";
 
 export default function RootLayout({
   children,
