+++ frontend/src/components/ui/NeedBadge.tsx
