--- frontend/src/constants/attendance.ts
@@ -1,9 +1,0 @@
-import { AttendanceFormData } from "@/types/attendance";
-
-export const EMPTY_FORM_ATTENDANCE: AttendanceFormData = {
-  studentId: "",
-  date: "",
-  type: "",
-  demand: "",
-  generalObservations: "",
-};
