--- frontend/src/constants/courses.ts
@@ -1,20 +1,0 @@
-import { CourseAcronym, CourseName } from "@/types/course";
-
-
-
-export const COURSES_ACRONYM: CourseAcronym[] = [
-  "E.S.",
-  "C.C.",
-  "A.B.I.",
-  "S.I.",
-  "Outros",
-];
-
-export const COURSES_NAME: CourseName[] = [
-  "Engenharia de Software",
-  "Ciência da Computação",
-  "A.B.I",
-  "Sistemas de Informação",
-  "Outros",
-];
-
