--- frontend/src/constants/student.ts
@@ -1,14 +1,0 @@
-import { StudentFormData } from "@/types/student";
-
-export const EMPTY_FORM: StudentFormData = {
-  externalId: "",
-  enrollmentId: "",
-  name: "",
-  dtBirth: "",
-  email: "",
-  phoneNumber: "",
-  courseId: "",
-  diagnosis: "",
-  potential: "",
-  difficulties: "",
-};
